MBU/backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Cors>=3.0
Flask-Bcrypt>=1.0
PyMySQL>=1.0 # Or mysql-connector-python if preferred
PyJWT>=2.0 # For authentication tokens
python-dotenv>=0.19
requests # If calling external APIs like Judge0 